Add Bootable USB toolchain
This commit contains the first working draft of a workflow to create a
bootable USB stick.

It is created by running base_image_util.sh NAME BOOL BOOL /dev/sd[x].

Its current functionality does not allow for the use of DMVerity and the
transperent block layer that verifies the rootfs during operation.