newscript: remove files we copy to; copy CONFIG to linux-stable/.config
If we are trying to copy a file to a place we need to remove the one that's
there. We remove it as it might be a symlink and opening and writing to it
will lead to some other file being written, which is bad.

We are once again copying CONFIG to linux-stable/.config; I forgot
that the make process in linux can change .config a lot.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>