Add a uinit for NiChrome
The uinit will take care of setting up all packages and doing
other housekeeping that's not really the job of xinit.

This is working (with the one simple change to the tcz list in usb).

We are going to reduce the u-root xinit to ... being an xinit, not
so many other things.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>