usb: move the tinycore usr and lib directories to cpio
The KERN-? partition is very tight on space. ROOT-? is quite
roomy. But we've been including usr/ and lib/ in KERN-?, even
though the usr/ and lib/
directories are needed for tinycore, not u-root.

Include usr/ and lib/ in tcz.cpio, and don't have them
included in the bb command. This greatly increases our
options as to what we include with u-root in the KERN-?
partition,.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>