install command
this install command is what you run to install NiChrome on a chromebook.
usage:
install <target-name>

For now it will always install on the B image

I've tested this on a chromebook and it works.

You do this:
boot the usb
install
remove usb and boot into chromeos
cgpt add -P 2 -T 1 -i 4 /dev/mmcblk1 # or whatever you device is)
reboot

and it goes to NiChrome, once. Next time you boot it's back to chromeos.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>