Support multiple user ids
This is now tested and we can run chrome as user 1000.

Just build the usb stick boot it, and you'll see three windows come up:
root window
user window
chrome (once you hit return for the ok)

This works but is still too slow. Next step is to prebuild some of the early startup binaries
so it's as fast as a busybox boot.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>