usb command now makes owner root:root and mirrors g perms to o
Sometimes users create directories such as tcz and user, and the
ownership ends up with a userid that makes no sense in nichrome.
Further, permissions of 750 are a problem when the default NiChrome
user (1000) starts up.

We can't come up with a case where we want to mirror user permissions
to the initramfs, so for now, set owner to root:root.

Further, mirror group settings to o, i.e. 750 goes to 755; 640 goes to 644,
and so on.

I tested this 3 ways.
First, I did not change the records, and the initramfs before and after
was the same.

Second, I added the owner and perm changes, and ownership went to root:root

Third, I set my top level tcz directory to 750, and observe that group
perms are mirrored to o perms in the cpio.

Remaining question: what should we do if a directory is 0700? I'm not sure.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>