WIP: allow command line selection of key type
Right now, we always select a standard key for a bootable stick.

This gross hack lets you do this
go run usb/*.go -keyfileprefix recovery_

and it will build a stick signed with the recovery key.

This is nasty, let's see if we even like it. Cherry-pick it, but don't
merge it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>