install: use a cpio passthrough function for root
Just doing a dumb raw dd of partitions is too dangerous.
This way we can figure out if the cpio is valid and not
copy the whole partition if the cpio is smaller (and it
is by a lot).

This does an install on my broken celes, I can't test it because
the celes is broken ... but it does do the full install in 18 seconds.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>