usb: require that /dev destinations be a device
We just wasted a day because the usb command doesn't require
that kernel and root paths in /dev be devices.

The real problem here is that /dev is a forty-seven-year-old idea that
should be gone and forgotten, but try to tell Unix users that.

Anyway ... if the dev path is in /dev, and is not /dev/null, then
require that the kern and root paths be devices.

Tested by creating, e.g., /dev/sdb3, and having it fail. You can still
create the kern and root as files, they just can't be in /dev.

Jorge, hope this helps.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>