uinit: add code to populate /home/user and .cpio
To get a reasonable startup screen we need to populate
/home/user/.config.

Permissions require that it be done by the initial user process,
not root; and that, further, it be done after the tmpfs mount
for /home/user.

This work can be done by uinit in the part of main that handles
user startup.

uinit will now look for /usr/user.cpio, as part of the user process
startup, and if it finds it it will unpack it.

note that the first user process now starts in /home/user, unlike
before, and because the cpio is done as that user, things will have
correct uid and permissions.

As a first step we add a cpio for chrome .config directories.

TODO: consider a generate step where we build the structure
of /home/user into uinit itself, in keeping with our goal of
functioning correctly even if files are missing.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>